Use LaserBladeVisual.Writer#writeModelType to change model type of laser blade

+++ src/main/java/com/github/iunius118/tolaserblade/core/laserblade/LaserBlade.java
@@ -101,7 +101,7 @@ public class LaserBlade {
         }
 
         public Writer writeType(int type) {
-            tag.putFloat(KEY_TYPE, Math.max(type, 0));
+            tag.putInt(KEY_TYPE, Math.max(type, 0));
             return this;
         }
     }
